Rename models.rs to query_plan.rs
