+++ crates/apollo-router-core/src/query_planner/mod.rs
@@ -1,9 +1,9 @@
 mod caching_query_planner;
-mod model;
+mod query_plan;
 mod router_bridge_query_planner;
 
 pub use caching_query_planner::*;
-pub use model::*;
+pub use query_plan::*;
 pub use router_bridge_query_planner::*;
 
 /// Query planning options.
